refactor(process-flow): StepCard.tsx の 15 kind body を sub-component に抽出 (#1145 Phase-2)
#1145 Phase-2 として、StepCard.tsx (76KB / 1689 行) の `step.kind === X` 別 inline JSX
body を 15 個の sub-component に分離。

- 各 kind の body を `internal/step-cards/<Kind>StepCardBody.tsx` に抽出
- branch / loop body の local UI state (`collapsedBranchIds` / `loopBodyCollapsed`)
  と関連 handler (`toggleBranchCollapse` / `setBranchAt` / `moveBranchUp/Down` /
  `deleteBranch` / `addBranch` / `addElseBranch`) を該当 sub-component に内部化
- 共通 props 型を `internal/step-cards/types.ts` に定義し、prop drilling の重複を抑制
- StepCard.tsx 1689 行 / 76KB -> 837 行 / 33KB (-50%, -43KB)
- 純粋なファイル分割で振る舞い変更なし

抽出した 15 kind: validation, dbAccess, externalSystem, commonProcess, compute,
return, screenTransition, displayUpdate, branch, loop, log, audit,
transactionScope, jump, workflow

残 22 - 15 = 7 kind (loopBreak/loopContinue/eventPublish/eventSubscribe/closing/
cdc/extension/other) は detail 専用 body を持たず metadata / notes のみで、
分割対象外。

Refs #1145
